+++ src/qow/framework/system/FrameLoop.java
@@ -7,7 +7,7 @@ import qow.framework.screen.QFrame;
  * 可能な限り一秒間に設定されたレートの回数の画面の再描写をする
  *
  * @author QOW
- * @version 2022/10/02
+ * @version 2023/04/04
  * @since 1.0.0
  */
 public class FrameLoop extends Loop {
@@ -34,7 +34,7 @@ public class FrameLoop extends Loop {
 
     /**
      * 可能な限り一秒間に設定されたレートの回数このメソッドを実行する<br>
-     * 指定時間内に処理が終わらない場合は{@link FrameLoop#overTime(double)}が呼び出される
+     * 指定時間内に処理が終わらない場合は{@link FrameLoop#overTime(long)}が呼び出される
      */
     public void loop() {
         rule.draw();
@@ -42,10 +42,11 @@ public class FrameLoop extends Loop {
 
     /**
      * 設定されたレートより時間がかかってしまった場合の処理
+     * 正の数値で返される
      *
-     * @param overTime 過ぎた時間のナノ秒
+     * @param overTime 過ぎた時間のミリ秒
      */
-    public void overTime(double overTime) {
+    public void overTime(long overTime) {
         rule.overTimeFrame(overTime);
     }
